refactor(web): transformed background on verify page BrickPhotoBackground.tsx component
