+++ apps/web/src/app/(account)/verify/user/[token]/page.tsx
@@ -1,42 +1,13 @@
 import React from "react"
 import { SetupAccountDialog } from "../_components/SetupAccountDialog";
 import { Toaster } from "@/components/ui/sonner"
-import Image from "next/image";
+import BrickPhotoBackground from "@/app/(account)/_components/BrickPhotoBackground";
 
 export default function Page() {
-    const photos = [
-        'https://images.unsplash.com/photo-1534260748473-e1c629d04bb0?q=80&w=2469&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1538329972958-465d6d2144ed?q=80&w=2970&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1589742090907-b64224d1a7e7?q=80&w=3087&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1634748210255-af588c16652a?q=80&w=3087&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1560253034-1a9c043ffb28?q=80&w=2970&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1651313950238-7f569c1fc2ba?q=80&w=3164&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1625697501049-ed422e855f4e?q=80&w=3135&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1529139574466-a303027c1d8b?q=80&w=3087&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1490114538077-0a7f8cb49891?q=80&w=2970&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1463100099107-aa0980c362e6?q=80&w=2970&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1651378499332-2d578930b9b4?q=80&w=3103&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1516700829972-65a01767c294?q=80&w=3174&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1511335413948-b5966e4e1d01?q=80&w=2973&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/flagged/photo-1553802922-2eb2f7f2c65b?q=80&w=2970&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-        'https://images.unsplash.com/photo-1588117305388-c2631a279f82?q=80&w=3087&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D'
-    ]
+    
     return (
         <div className="h-screen flex justify-center items-center relative overflow-hidden">
-            <div className="
-                absolute md:columns-4 md:space-x-2 md:space-y-6 md:p-2 max-sm:hidden
-                md:rotate-[20deg] md:w-[145%] bg-black
-                sm:columns-3 sm:space-x-1 sm:space-y-4 sm:p-4 duration-200 lg:columns-5
-            ">
-                {
-                    photos.map((photo, idx) => {
-                        return <div className="rounded-xl overflow-hidden" key={idx}>
-                            <Image width={500} height={500} src={photo} alt="" />
-                        </div>
-                    })
-                }
-
-            </div>
+            <BrickPhotoBackground />
             <span className="absolute bottom-4 right-[5%] font-thin text-base text-white items-start hidden lg:flex">WearDrobe <span className="text-xs">&copy;</span></span>
             <SetupAccountDialog className="absolute" title={"Just a few more steps.."}  />
             <Toaster position="top-center" richColors />
